+++ app/halgo/Halgo/Util.hs
@@ -26,7 +26,7 @@ import Network.HTTP.Types (Status (statusCode, statusMessage))
 import Servant.Client (ClientError (..), ResponseF (..))
 import Servant.Client.Generic (AsClientT)
 
-import Network.Algorand.Api (ApiIdx2, ApiV2)
+import Network.Algorand.Api (ApiV2, IndexerApi)
 import Network.Algorand.Client (AlgoIndexer (..), AlgoNode (..), connectToIndexer, connectToNode)
 import Network.Algorand.Definitions (Host, Network)
 
@@ -73,7 +73,7 @@ withNode url act = do
 withIndexer
   :: MonadSubCommand m
   => Host
-  -> ((Network, ApiIdx2 (AsClientT m)) -> m a)
+  -> ((Network, IndexerApi (AsClientT m)) -> m a)
   -> m a
 withIndexer url act = do
   net <- asks goNetwork
